connect payment to the order_id and fix

+++ aviation/payments/models.py
@@ -6,9 +6,9 @@ class Payment(models.Model):
     order = models.ForeignKey("orders.Order", on_delete=models.CASCADE, null=True, blank=True)
     stripe_session_id = models.CharField(max_length=255, unique=True)
     stripe_payment_intent = models.CharField(max_length=255, blank=True, null=True)
-    amount = models.IntegerField()  # у центах
+    amount = models.DecimalField(max_digits=10, decimal_places=2)
     currency = models.CharField(max_length=10, default="usd")
-    status = models.CharField(max_length=50, default="pending")  # pending/paid/failed
+    status = models.CharField(max_length=50, default="pending")
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
